requirements.txt
python-amazon-sp-api==0.16.0
pandas==1.4.3
